fixing watcher to work with app module changes
