+++ src/render_engine/watcher/__init__.py
@@ -1,1 +1,1 @@
-from .event import RegExHandler
+from .event import RegExHandler, get_app
